# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte ETV",
        "description": "REST API. Projecte ETV. DAW Client i servidor.",
        "contact": {
            "name": "Webmaster.",
            "email": "info@webmaster.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.dawpaucasesnoves.com/etvServidor/public"
        }
    ],
    "paths": {
...
        "/api/zones": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Mostrar totes les zones",
                "description": "Display a listing of the resource.",
                "operationId": "9c05a68fa58d9de55bf6e8ce07a41f2d",
                "responses": {
                    "200": {
                        "description": "Mostrar totes les zones.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zones/{id}": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Mostrar una zona",
                "description": "Store a newly created resource in storage.",
                "operationId": "64815812122d48340afbf2f8e3e0e1fb",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "missatge": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string"
                                        },
                                        "codi": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}